Fix unwinding from top-left corner (#45)
